# backend/docs/swagger.yaml
basePath: /
definitions:
...
  v1.ErrorResponse:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
...
  v1.TrackPageResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/v1.TrackResponse'
        type: array
      limit:
        type: integer
      next_cursor:
        type: string
    type: object
  v1.TrackPatchRequest:
    properties:
      artist:
        type: string
      title:
        type: string
    type: object
  v1.TrackResponse:
    properties:
      artist:
        type: string
      id:
        type: string
      size_bytes:
        type: integer
      status:
        type: string
      title:
        type: string
      user_id:
        type: string
    type: object
  v1.TrackUploadInitRequest:
    properties:
      artist:
        type: string
      size:
        type: integer
      title:
        type: string
    type: object
  v1.TrackUploadInitResponse:
    properties:
      track:
        $ref: '#/definitions/v1.TrackResponse'
      upload_url:
        type: string
    type: object
  v1.TrackWriteRequest:
    properties:
      artist:
        type: string
      title:
        type: string
    type: object
...
info:
  contact: {}
  description: Music streaming backend
  title: Chimera API
  version: "1.0"
paths:
...
  /v1/me/likes:
    get:
      operationId: listMyLikes
      parameters:
      - default: 20
        description: Page size
        in: query
        name: limit
        type: integer
      - description: Pagination cursor
        in: query
        name: cursor
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.TrackPageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      security:
      - BearerAuth: []
      summary: My liked tracks
      tags:
      - tracks
...
  /v1/me/tracks:
    get:
      description: All tracks uploaded by the current user, including drafts.
      operationId: listMyTracks
      parameters:
      - default: 20
        description: Page size
        in: query
        name: limit
        type: integer
      - description: Pagination cursor
        in: query
        name: cursor
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.TrackPageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      security:
      - BearerAuth: []
      summary: My uploaded tracks
      tags:
      - tracks
...
  /v1/tracks:
    get:
      description: Lists ready tracks only.
      operationId: listTracks
      parameters:
      - description: Exact artist credit (case-insensitive)
        in: query
        name: artist
        type: string
      - default: 20
        description: Page size
        in: query
        name: limit
        type: integer
      - description: Pagination cursor
        in: query
        name: cursor
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.TrackPageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: Track feed
      tags:
      - tracks
  /v1/tracks/{trackId}:
    delete:
      operationId: deleteTrack
      parameters:
      - description: Track ID
        in: path
        name: trackId
        required: true
        type: string
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete track
      tags:
      - tracks
    get:
      operationId: getTrack
      parameters:
      - description: Track ID
        in: path
        name: trackId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.TrackResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: Get track by id
      tags:
      - tracks
    patch:
      consumes:
      - application/json
      operationId: updateTrack
      parameters:
      - description: Track ID
        in: path
        name: trackId
        required: true
        type: string
      - description: Track fields
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/v1.TrackPatchRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.TrackResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update track fields
      tags:
      - tracks
    put:
      consumes:
      - application/json
      operationId: replaceTrack
      parameters:
      - description: Track ID
        in: path
        name: trackId
        required: true
        type: string
      - description: Track
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/v1.TrackWriteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.TrackResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update track
      tags:
      - tracks
  /v1/tracks/{trackId}/like:
    delete:
      operationId: unlikeTrack
      parameters:
      - description: Track ID
        in: path
        name: trackId
        required: true
        type: string
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Remove a like
      tags:
      - tracks
    post:
      operationId: likeTrack
      parameters:
      - description: Track ID
        in: path
        name: trackId
        required: true
        type: string
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Like a ready track
      tags:
      - tracks
  /v1/tracks/{trackId}/stream:
    get:
      operationId: streamTrack
      parameters:
      - description: Track ID
        in: path
        name: trackId
        required: true
        type: string
      responses:
        "302":
          description: Redirect to audio
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: Stream a ready track
      tags:
      - tracks
  /v1/tracks/{trackId}/upload-complete:
    post:
      description: Checks the object in MinIO and publishes the track as ready.
      operationId: completeTrackUpload
      parameters:
      - description: Track ID
        in: path
        name: trackId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.TrackResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Finish track upload
      tags:
      - tracks
  /v1/tracks/upload-init:
    post:
      consumes:
      - application/json
      operationId: initTrackUpload
      parameters:
      - description: Track metadata
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/v1.TrackUploadInitRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/v1.TrackUploadInitResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Start track upload
      tags:
      - tracks
...
  /v1/users/{userId}/tracks:
    get:
      description: Ready tracks uploaded by the given user.
      operationId: listUserTracks
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      - default: 20
        description: Page size
        in: query
        name: limit
        type: integer
      - description: Pagination cursor
        in: query
        name: cursor
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.TrackPageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: Tracks uploaded by a user
      tags:
      - tracks
securityDefinitions:
  BearerAuth:
    description: JWT from /v1/auth/login. Paste only the token, or "Bearer <token>".
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"